Resolviendo accesibilidad de Usuario

+++ backend/src/main/java/com/reforestart/backend/dto/ArbolDTO.java
@@ -16,6 +16,6 @@ public class ArbolDTO {
 
     private String descripcion;
 
-    private Long precio;
+    private Integer precio;
 
 }
